Added [DefaultExecutionOrder(-50)] to MonoSingleton
Afaik, this causes no harm and ensures the singleton functions(such as
Awake, Update etc.) are executed before other classes.
This can be changed in inherited classes if needed, but in 99% of the
cases, I would say that you'd want the singleton to execute first.
-50 seemed like a reasonable choice due to certain built in classes
utilizing up to -100(e.g. UnityEngine.InputSystem.PlayerInput) and this
way we still get plenty of free room around the singleton.